Functional re-implementation of the original naive console.
Fixed two bugs in the PanawaveStruct.add_ring method which arose in
testing console input:
- evaluated args were appended twice
- kwargs were not passed on to StickerRing.__init__()

The layout of the console button and entry widgets is still broken but
generally functional. Rather than developing this any further or
attempting to implement IPython embedding within a tkinter widget (which
is non-trivial), will probably leave this as-is once the layout is
corrected, and add an additional widget which spawns a full IPython
Embed within an xterm, once the existing debugging embed is removed.